# tools/schemas/openapi_specification.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smelter",
    "description": "Real-time video compositing software",
    "license": {
      "name": "MIT"
    },
    "version": "0.6.0"
  },
  "paths": {
...
    "/api/input/{input_id}/register": {
      "post": {
        "tags": [
          "register_request"
        ],
        "operationId": "register_input",
        "parameters": [
          {
            "name": "input_id",
            "in": "path",
            "description": "Input ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Input registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterInputResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/output/{output_id}/register": {
      "post": {
        "tags": [
          "register_request"
        ],
        "operationId": "register_output",
        "parameters": [
          {
            "name": "output_id",
            "in": "path",
            "description": "Output ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterOutput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Output registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterOutputResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/shader/{shader_id}/register": {
      "post": {
        "tags": [
          "register_request"
        ],
        "operationId": "register_shader",
        "parameters": [
          {
            "name": "shader_id",
            "in": "path",
            "description": "Shader ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShaderSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Shader registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/web-renderer/{instance_id}/register": {
      "post": {
        "tags": [
          "register_request"
        ],
        "operationId": "register_web_renderer",
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "description": "Web renderer instance ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebRendererSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Web renderer registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/image/{image_id}/register": {
      "post": {
        "tags": [
          "register_request"
        ],
        "operationId": "register_image",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "description": "Image ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Image registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/font/register": {
      "post": {
        "tags": [
          "register_request"
        ],
        "operationId": "register_font",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/RegisterFontRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Font registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": [
          "error_code",
          "message",
          "stack",
          "http_status_code"
        ],
        "properties": {
          "error_code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "stack": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "http_status_code": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "ImageSpec": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "asset_type"
            ],
            "properties": {
              "url": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "path": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "asset_type": {
                "type": "string",
                "enum": [
                  "png"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "asset_type"
            ],
            "properties": {
              "url": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "path": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "asset_type": {
                "type": "string",
                "enum": [
                  "jpeg"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "asset_type"
            ],
            "properties": {
              "url": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "path": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "resolution": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/Resolution"
                  }
                ]
              },
              "asset_type": {
                "type": "string",
                "enum": [
                  "svg"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "asset_type"
            ],
            "properties": {
              "url": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "path": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "asset_type": {
                "type": "string",
                "enum": [
                  "gif"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "asset_type"
            ],
            "properties": {
              "url": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "path": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "asset_type": {
                "type": "string",
                "enum": [
                  "auto"
                ]
              }
            }
          }
        ]
      },
...
      "OkResponse": {
        "type": "object",
        "description": "Response for requests that do not return any data."
      },
...
      "RegisterFontRequest": {
        "type": "object",
        "required": [
          "file"
        ],
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "contentMediaType": "application/octet-stream"
          }
        }
      },
      "RegisterInput": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RtpInput"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "rtp_stream"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RtmpInput"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "rtmp_server"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/MoqServerInput"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "moq_server"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/MoqClientInput"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "moq_client"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/Mp4Input"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "mp4"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/WhipInput"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "whip_server"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/WhepInput"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "whep_client"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/HlsInput"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "hls"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/V4l2Input"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "v4l2"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/DeckLink"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "decklink"
                    ]
                  }
                }
              }
            ]
          }
        ]
      },
      "RegisterInputResponse": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "port": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int32",
                "minimum": 0
              }
            }
          },
          {
            "type": "object",
            "properties": {
              "video_duration_ms": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int64",
                "minimum": 0
              },
              "audio_duration_ms": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int64",
                "minimum": 0
              }
            }
          },
          {
            "type": "object",
            "required": [
              "bearer_token",
              "endpoint_route"
            ],
            "properties": {
              "bearer_token": {
                "type": "string"
              },
              "endpoint_route": {
                "type": "string"
              }
            }
          },
          {
            "type": "object"
          }
        ]
      },
      "RegisterOutput": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RtpOutput"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "rtp_stream"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RtmpOutput"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "rtmp_client"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/MoqClientOutput"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "moq_client"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/Mp4Output"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "mp4"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/WhipOutput"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "whip_client"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/WhepOutput"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "whep_server"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/HlsOutput"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "hls"
                    ]
                  }
                }
              }
            ]
          }
        ]
      },
      "RegisterOutputResponse": {
        "type": "object",
        "properties": {
          "port": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Port allocated for the output. Only returned for RTP outputs.",
            "minimum": 0
          }
        }
      },
...
      "Resolution": {
        "type": "object",
        "required": [
          "width",
          "height"
        ],
        "properties": {
          "width": {
            "type": "integer",
            "description": "Width in pixels.",
            "minimum": 0
          },
          "height": {
            "type": "integer",
            "description": "Height in pixels.",
            "minimum": 0
          }
        }
      },
...
      "RtmpOutput": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string",
            "description": "RTMP endpoint url."
          },
          "video": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OutputRtmpClientVideoOptions",
                "description": "Video stream configuration."
              }
            ]
          },
          "audio": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OutputRtmpClientAudioOptions",
                "description": "Audio stream configuration."
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "RtpInput": {
        "type": "object",
        "description": "Parameters for an input stream from RTP source.\nAt least one of `video` and `audio` has to be defined.",
        "required": [
          "port"
        ],
        "properties": {
          "port": {
            "$ref": "#/components/schemas/PortOrPortRange",
            "description": "UDP port or port range on which the compositor should listen for the stream."
          },
          "transport_protocol": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TransportProtocol",
                "description": "Transport protocol."
              }
            ]
          },
          "video": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/InputRtpVideoOptions",
                "description": "Parameters of a video source included in the RTP stream."
              }
            ]
          },
          "audio": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/InputRtpAudioOptions",
                "description": "Parameters of an audio source included in the RTP stream."
              }
            ]
          },
          "required": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "If input is required and the stream is not delivered on time, then Smelter will delay\nproducing output frames.\n\nDefaults to `false`."
          },
          "offset_ms": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Offset in milliseconds relative to the pipeline start (start request). If the offset is\nnot defined then the stream will be synchronized based on the delivery time of the initial\nframes."
          },
          "buffer_size_ms": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Size of the jitter buffer in milliseconds. Controls how long packets are held to\nabsorb network jitter and reorder out-of-order packets. Higher values increase\nlatency but improve resilience to packet loss and reordering."
          },
          "side_channel": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SideChannel",
                "description": "Enable side channel for video and/or audio track."
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "ShaderSpec": {
        "type": "object",
        "required": [
          "source"
        ],
        "properties": {
          "source": {
            "type": "string",
            "description": "Shader source code."
          }
        },
        "additionalProperties": false
      },
...
      "WebRendererSpec": {
        "type": "object",
        "required": [
          "url",
          "resolution"
        ],
        "properties": {
          "url": {
            "type": "string",
            "description": "Url of a website that you want to render."
          },
          "resolution": {
            "$ref": "#/components/schemas/Resolution",
            "description": "Resolution."
          },
          "embedding_method": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WebEmbeddingMethod",
                "description": "Mechanism used to render input frames on the website."
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}